+++ modules/AWSPowerShell/Cmdlets/AmazonElasticBeanstalk/Basic/Get-EBEnvironmentHealth-Cmdlet.cs
@@ -1,0 +1,151 @@
+/*******************************************************************************
+ *  Copyright 2012-2015 Amazon.com, Inc. or its affiliates. All Rights Reserved.
+ *  Licensed under the Apache License, Version 2.0 (the "License"). You may not use
+ *  this file except in compliance with the License. A copy of the License is located at
+ *
+ *  http://aws.amazon.com/apache2.0
+ *
+ *  or in the "license" file accompanying this file.
+ *  This file is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR
+ *  CONDITIONS OF ANY KIND, either express or implied. See the License for the
+ *  specific language governing permissions and limitations under the License.
+ * *****************************************************************************
+ *
+ *  AWS Tools for Windows (TM) PowerShell (TM)
+ *
+ */
+
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Management.Automation;
+using System.Text;
+using Amazon.PowerShell.Common;
+using Amazon.Runtime;
+using Amazon.ElasticBeanstalk;
+using Amazon.ElasticBeanstalk.Model;
+
+namespace Amazon.PowerShell.Cmdlets.EB
+{
+    /// <summary>
+    /// Returns information about the overall health of the specified environment. The <b>DescribeEnvironmentHealth</b>
+    /// operation is only available with AWS Elastic Beanstalk Enhanced Health.
+    /// </summary>
+    [Cmdlet("Get", "EBEnvironmentHealth")]
+    [OutputType("Amazon.ElasticBeanstalk.Model.DescribeEnvironmentHealthResponse")]
+    [AWSCmdlet("Invokes the DescribeEnvironmentHealth operation against AWS Elastic Beanstalk.", Operation = new[] {"DescribeEnvironmentHealth"})]
+    [AWSCmdletOutput("Amazon.ElasticBeanstalk.Model.DescribeEnvironmentHealthResponse",
+        "This cmdlet returns a DescribeEnvironmentHealthResponse object containing multiple properties. The object can also be referenced from properties attached to the cmdlet entry in the $AWSHistory stack."
+    )]
+    public class GetEBEnvironmentHealthCmdlet : AmazonElasticBeanstalkClientCmdlet, IExecutor
+    {
+        /// <summary>
+        /// <para>
+        /// <para>Specifies the response elements you wish to receive. If no attribute names are specified,
+        /// AWS Elastic Beanstalk returns all response elements.</para>
+        /// </para>
+        /// </summary>
+        [System.Management.Automation.Parameter]
+        [Alias("AttributeNames")]
+        public System.String[] AttributeName { get; set; }
+        
+        /// <summary>
+        /// <para>
+        /// <para>Specifies the AWS Elastic Beanstalk environment ID.</para>
+        /// </para>
+        /// </summary>
+        [System.Management.Automation.Parameter(Position = 0, ValueFromPipelineByPropertyName = true, ValueFromPipeline = true)]
+        public String EnvironmentId { get; set; }
+        
+        /// <summary>
+        /// <para>
+        /// <para>Specifies the AWS Elastic Beanstalk environment name.</para>
+        /// </para>
+        /// </summary>
+        [System.Management.Automation.Parameter(ValueFromPipelineByPropertyName = true)]
+        public String EnvironmentName { get; set; }
+        
+        
+        protected override void ProcessRecord()
+        {
+            base.ProcessRecord();
+            
+            var context = new CmdletContext
+            {
+                Region = this.Region,
+                Credentials = this.CurrentCredentials
+            };
+            
+            if (this.AttributeName != null)
+            {
+                context.AttributeNames = new List<String>(this.AttributeName);
+            }
+            context.EnvironmentId = this.EnvironmentId;
+            context.EnvironmentName = this.EnvironmentName;
+            
+            var output = Execute(context) as CmdletOutput;
+            ProcessOutput(output);
+        }
+        
+        #region IExecutor Members
+        
+        public object Execute(ExecutorContext context)
+        {
+            var cmdletContext = context as CmdletContext;
+            // create request
+            var request = new DescribeEnvironmentHealthRequest();
+            
+            if (cmdletContext.AttributeNames != null)
+            {
+                request.AttributeNames = cmdletContext.AttributeNames;
+            }
+            if (cmdletContext.EnvironmentId != null)
+            {
+                request.EnvironmentId = cmdletContext.EnvironmentId;
+            }
+            if (cmdletContext.EnvironmentName != null)
+            {
+                request.EnvironmentName = cmdletContext.EnvironmentName;
+            }
+            
+            CmdletOutput output;
+            
+            // issue call
+            var client = Client ?? CreateClient(context.Credentials, context.Region);
+            try
+            {
+                var response = client.DescribeEnvironmentHealth(request);
+                Dictionary<string, object> notes = null;
+                object pipelineOutput = response;
+                output = new CmdletOutput
+                {
+                    PipelineOutput = pipelineOutput,
+                    ServiceResponse = response,
+                    Notes = notes
+                };
+            }
+            catch (Exception e)
+            {
+                output = new CmdletOutput { ErrorResponse = e };
+            }
+            
+            return output;
+        }
+        
+        public ExecutorContext CreateContext()
+        {
+            return new CmdletContext();
+        }
+        
+        #endregion
+        
+        
+        internal class CmdletContext : ExecutorContext
+        {
+            public List<String> AttributeNames { get; set; }
+            public String EnvironmentId { get; set; }
+            public String EnvironmentName { get; set; }
+        }
+        
+    }
+}